Refactor fuels-programs/contract
