+++ packages/fuels-programs/src/calls.rs
@@ -1,10 +1,12 @@
-mod contract;
+mod call_handler;
+mod contract_call;
 pub mod receipt_parser;
-mod script;
+mod script_call;
 pub mod traits;
 mod tx_dependency_extension;
 pub mod utils;
 
-pub use contract::*;
-pub use script::*;
+pub use call_handler::*;
+pub use contract_call::*;
+pub use script_call::*;
 pub use tx_dependency_extension::*;
