+++ packages/fuels-programs/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod call_handler;
 pub mod calls;
 pub mod contract;
 pub mod responses;
